requirements.txt
# requirements.txt
sympy
pandas
openpyxl
numpy
scipy
python-docx